+++ core/src/main/java/audios/EfectoSonido.java
@@ -4,18 +4,11 @@ import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.audio.Sound;
 
 public class EfectoSonido {
-    private Sound efecto;
-
-    public EfectoSonido(String nombreArchivo) {
-        efecto = Gdx.audio.newSound(Gdx.files.internal("sonidos/" + nombreArchivo + ".mp3"));
-    }
-
-    public void reproducir() {
-        efecto.play();
-    }
-
-    public void reproducir(float volumen) {
-        efecto.play(volumen);
+    private static Sound efecto;
+    
+    public static void reproducir(String nombreArchivo, float volumen) {
+    	efecto = Gdx.audio.newSound(Gdx.files.internal("sonidos/" + nombreArchivo + ".mp3"));
+    	efecto.play(volumen);
     }
 
     public void liberar() {