+++ core/src/main/java/personajes/Akame.java
@@ -17,27 +17,27 @@ public class Akame extends Personaje{
 	        for (int i = 1; i <= 4; i++) {
 	            framesDerecha.add(new TextureRegion(new Texture(Gdx.files.internal("imagenes/personajes/akame/akame_derecha_moviendose_" + i + ".png"))));
 	        }
-	        animDerecha = new Animation<>(0.1f, framesDerecha, Animation.PlayMode.LOOP);
+	        super.animDerecha = new Animation<>(0.1f, framesDerecha, Animation.PlayMode.LOOP);
 	        
 	        Array<TextureRegion> framesAtaqueDerecha = new Array<>();
 	        for (int i = 1; i <= 6; i++) {
 	        	framesAtaqueDerecha.add(new TextureRegion(new Texture(Gdx.files.internal("imagenes/personajes/akame/ataque/akame_derecha_atacando_" + i + ".png"))));
 	        }
-	        animAtaqueDerecha = new Animation<>(0.1f, framesAtaqueDerecha, Animation.PlayMode.LOOP);
+	        super.animAtaqueDerecha = new Animation<>(0.1f, framesAtaqueDerecha, Animation.PlayMode.LOOP);
 	        
 	        Array<TextureRegion> framesAtaqueIzquierda = new Array<>();
 	        for (int i = 1; i <= 6; i++) {
 	        	framesAtaqueIzquierda.add(new TextureRegion(new Texture(Gdx.files.internal("imagenes/personajes/akame/ataque/akame_izquierda_atacando_" + i + ".png"))));
 	        }
-	        animAtaqueIzquierda = new Animation<>(0.1f, framesAtaqueIzquierda, Animation.PlayMode.LOOP);
+	        super.animAtaqueIzquierda = new Animation<>(0.1f, framesAtaqueIzquierda, Animation.PlayMode.LOOP);
 
 	        Array<TextureRegion> framesIzquierda = new Array<>();
 	        for (int i = 1; i <= 4; i++) {
 	            framesIzquierda.add(new TextureRegion(new Texture(Gdx.files.internal("imagenes/personajes/akame/akame_izquierda_moviendose_" + i + ".png"))));
 	        }
-	        animIzquierda = new Animation<>(0.1f, framesIzquierda, Animation.PlayMode.LOOP);
+	        super.animIzquierda = new Animation<>(0.1f, framesIzquierda, Animation.PlayMode.LOOP);
 
-	        quietaDerecha = new TextureRegion(new Texture(Gdx.files.internal("imagenes/personajes/akame/akame_derecha_(detenida).png")));
-	        quietaIzquierda = new TextureRegion(new Texture(Gdx.files.internal("imagenes/personajes/akame/akame_izquierda_(detenida).png")));
+	        super.quietaDerecha = new TextureRegion(new Texture(Gdx.files.internal("imagenes/personajes/akame/akame_derecha_(detenida).png")));
+	        super.quietaIzquierda = new TextureRegion(new Texture(Gdx.files.internal("imagenes/personajes/akame/akame_izquierda_(detenida).png")));
 	    }
 }