+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [4.5.9] - 2025-08-16
+### Modificado
+- El volumen del sonido de ataque ahora esta compartido con el de la musica, de modo que cuando el volumen de musica esta bajo, el del sonido tambien. La musica de derrota tambien. 
+- Optimizacion del metodo atacar. 
+
 ## [4.5.8] - 2025-08-04
 ### Modificado
 - La colision del objeto desaparece junto a este